[backend] Refact tests
