+++ openbas-api/src/test/java/io/openbas/rest/AssetGroupApiTest.java
@@ -6,8 +6,8 @@ import io.openbas.database.model.AssetGroup;
 import io.openbas.database.repository.AssetGroupRepository;
 import io.openbas.database.repository.EndpointRepository;
 import io.openbas.rest.asset_group.form.AssetGroupInput;
-import io.openbas.rest.utils.WithMockObserverUser;
-import io.openbas.rest.utils.WithMockPlannerUser;
+import io.openbas.utils.mockUser.WithMockObserverUser;
+import io.openbas.utils.mockUser.WithMockPlannerUser;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,8 +17,8 @@ import org.springframework.test.web.servlet.MockMvc;
 import java.util.List;
 
 import static io.openbas.rest.asset_group.AssetGroupApi.ASSET_GROUP_URI;
-import static io.openbas.rest.utils.JsonUtils.asJsonString;
-import static io.openbas.rest.utils.JsonUtils.asStringJson;
+import static io.openbas.utils.JsonUtils.asJsonString;
+import static io.openbas.utils.JsonUtils.asStringJson;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS;
