+++ openbas-api/src/test/java/io/openbas/rest/EndpointApiTest.java
@@ -5,8 +5,8 @@ import io.openbas.IntegrationTest;
 import io.openbas.database.model.Endpoint;
 import io.openbas.database.repository.EndpointRepository;
 import io.openbas.rest.asset.endpoint.form.EndpointInput;
-import io.openbas.rest.utils.WithMockObserverUser;
-import io.openbas.rest.utils.WithMockPlannerUser;
+import io.openbas.utils.mockUser.WithMockObserverUser;
+import io.openbas.utils.mockUser.WithMockPlannerUser;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,8 +17,8 @@ import java.util.List;
 
 import static io.openbas.database.model.Endpoint.PLATFORM_TYPE.Linux;
 import static io.openbas.rest.asset.endpoint.EndpointApi.ENDPOINT_URI;
-import static io.openbas.rest.utils.JsonUtils.asJsonString;
-import static io.openbas.rest.utils.JsonUtils.asStringJson;
+import static io.openbas.utils.JsonUtils.asJsonString;
+import static io.openbas.utils.JsonUtils.asStringJson;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
