+++ openbas-api/src/test/java/io/openbas/rest/InjectApiTest.java
@@ -6,8 +6,8 @@ import io.openbas.database.model.Scenario;
 import io.openbas.database.repository.InjectRepository;
 import io.openbas.database.repository.ScenarioRepository;
 import io.openbas.rest.inject.form.InjectInput;
-import io.openbas.rest.utils.WithMockObserverUser;
-import io.openbas.rest.utils.WithMockPlannerUser;
+import io.openbas.utils.mockUser.WithMockObserverUser;
+import io.openbas.utils.mockUser.WithMockPlannerUser;
 import io.openbas.service.ScenarioService;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +18,7 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import static io.openbas.injects.email.EmailContract.EMAIL_DEFAULT;
 import static io.openbas.rest.scenario.ScenarioApi.SCENARIO_URI;
-import static io.openbas.rest.utils.JsonUtils.asJsonString;
+import static io.openbas.utils.JsonUtils.asJsonString;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS;
