+++ openbas-api/src/test/java/io/openbas/rest/ScenarioApiTest.java
@@ -4,8 +4,8 @@ import com.jayway.jsonpath.JsonPath;
 import io.openbas.database.repository.ScenarioRepository;
 import io.openbas.rest.scenario.form.ScenarioInformationInput;
 import io.openbas.rest.scenario.form.ScenarioInput;
-import io.openbas.rest.utils.WithMockObserverUser;
-import io.openbas.rest.utils.WithMockPlannerUser;
+import io.openbas.utils.mockUser.WithMockObserverUser;
+import io.openbas.utils.mockUser.WithMockPlannerUser;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -14,7 +14,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static io.openbas.rest.scenario.ScenarioApi.SCENARIO_URI;
-import static io.openbas.rest.utils.JsonUtils.asJsonString;
+import static io.openbas.utils.JsonUtils.asJsonString;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS;
