+++ openbas-api/src/test/java/io/openbas/rest/UserApiTest.java
@@ -5,15 +5,15 @@ import io.openbas.database.model.User;
 import io.openbas.database.repository.UserRepository;
 import io.openbas.rest.user.form.login.LoginUserInput;
 import io.openbas.rest.user.form.user.CreateUserInput;
-import io.openbas.rest.utils.fixtures.UserFixture;
+import io.openbas.utils.fixtures.UserFixture;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 
-import static io.openbas.rest.utils.JsonUtils.asJsonString;
-import static io.openbas.rest.utils.fixtures.UserFixture.EMAIL;
+import static io.openbas.utils.JsonUtils.asJsonString;
+import static io.openbas.utils.fixtures.UserFixture.EMAIL;
 import static org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
