+++ openbas-api/src/test/java/io/openbas/utils/fixtures/ArticleFixture.java
@@ -1,4 +1,4 @@
-package io.openbas.rest.utils.fixtures;
+package io.openbas.utils.fixtures;
 
 import io.openbas.database.model.Article;
 import io.openbas.database.model.Channel;
