+++ openbas-api/src/test/java/io/openbas/utils/fixtures/DocumentFixture.java
@@ -1,4 +1,4 @@
-package io.openbas.rest.utils.fixtures;
+package io.openbas.utils.fixtures;
 
 import io.openbas.database.model.Document;
