+++ openbas-api/src/test/java/io/openbas/utils/fixtures/UserFixture.java
@@ -1,4 +1,4 @@
-package io.openbas.rest.utils.fixtures;
+package io.openbas.utils.fixtures;
 
 import io.openbas.database.model.User;
 import io.openbas.rest.user.form.login.LoginUserInput;
