+++ openbas-api/src/test/java/io/openbas/utils/mockUser/WithMockAdminUser.java
@@ -1,11 +1,11 @@
-package io.openbas.utils;
+package io.openbas.utils.mockUser;
 
 import org.springframework.security.test.context.support.WithSecurityContext;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
-import static io.openbas.utils.WithMockAdminUserSecurityContextFactory.MOCK_USER_ADMIN_EMAIL;
+import static io.openbas.utils.mockUser.WithMockAdminUserSecurityContextFactory.MOCK_USER_ADMIN_EMAIL;
 
 @Retention(RetentionPolicy.RUNTIME)
 @WithSecurityContext(factory = WithMockAdminUserSecurityContextFactory.class)
