+++ openbas-api/src/test/java/io/openbas/utils/mockUser/WithMockPlannerUser.java
@@ -1,11 +1,11 @@
-package io.openbas.rest.utils;
+package io.openbas.utils.mockUser;
 
 import org.springframework.security.test.context.support.WithSecurityContext;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
-import static io.openbas.rest.utils.WithMockPlannerUserSecurityContextFactory.MOCK_USER_PLANNER_EMAIL;
+import static io.openbas.utils.mockUser.WithMockPlannerUserSecurityContextFactory.MOCK_USER_PLANNER_EMAIL;
 
 @Retention(RetentionPolicy.RUNTIME)
 @WithSecurityContext(factory = WithMockPlannerUserSecurityContextFactory.class)
